// proj_monitoramento/backend/src/modules/user/Interfaces/ICreateUserFiscalDTO.ts
interface ICreateUserFiscalDTO {
 
  nome_fiscal: string; 
  nome_orgao: string;
  cargo: string;
  profissao: string;
  nivel_escolaridade: string;
  sexo: string;
  genero: string;
  cpf_fiscal: string; 
  mat_orgao: string;
  unidade_sei: string;
  email_institucional: string;
  email_pessoal: string;
  cnpj_fiscal: string;
  nome_empresa: string;
  num_contrato: string;
  password: string;
  

  }
  
  export { ICreateUserFiscalDTO };
  